ajout contrainte sur choix du temps pour un TS

+++ resources/js/helpers/timesheetHelper.js
@@ -28,10 +28,10 @@ export function getTimesheetsConsumption(timesheets) {
         return { totalWorktime: 0, totalWorktimeUpdatedOnPodio: 0 };
     }
     const totalWorktime = timesheets.reduce((currentTotal, item) => {
-        return parseInt(item.worktime) + currentTotal;
+        return parseFloat(item.worktime) + currentTotal;
     }, 0);
     const totalWorktimeUpdatedOnPodio = timesheets.reduce((currentTotal, item) => {
-        return item.is_podio_updated ? parseInt(item.worktime) + currentTotal : currentTotal;
+        return item.is_podio_updated ? parseFloat(item.worktime) + currentTotal : currentTotal;
     }, 0);
 
     return { totalWorktime, totalWorktimeUpdatedOnPodio };
